MOVE transport time series code
My first check in of the code to calculate the transport for the MOVE region.  This code calculates the transport using model velocities, and geostrophic velocities using a top down methodology and a bottom up methodology.  It calculates the geostrophic velocities using all available columns and only the SALT, TEMP and SSH at the two boundaries at each side.